requirements.txt
Django
djangorestframework
python-decouple
pillow
psycopg2-binary
django-jazzmin
drf_yasg
channels
